Fixes incorrect progress when installing multiple apps in single app session
We incorrectly assumed @TaskLocal default values are created per-task, when instead they are created only once and then reused as default. This is fine for value types, but for reference types it can result in weird bugs.

To fix this, we now explicitly create a new Progress (and OperationContext) when installing or updating an app.